feat: add demo class
- Move main function into demo to keep main clean.
- Add one additional unit test for demo

Closes #71 , closes #61
